Move mapping code in a MappingResolver class
