+++ tests/FlatMapperCreateMappingTest.php
@@ -9,6 +9,7 @@ use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
 use Pixelshaped\FlatMapperBundle\Exception\MappingCreationException;
 use Pixelshaped\FlatMapperBundle\FlatMapper;
+use Pixelshaped\FlatMapperBundle\MappingResolver;
 use Pixelshaped\FlatMapperBundle\Tests\Examples\Invalid\Circular\CycleRootDTO;
 use Pixelshaped\FlatMapperBundle\Tests\Examples\Invalid\NameTransformation\InvalidNameTransformationDTO;
 use Pixelshaped\FlatMapperBundle\Tests\Examples\Invalid\RootAbstractDTO;
@@ -34,11 +35,12 @@ use Symfony\Contracts\Cache\CacheInterface;
 use Symfony\Contracts\Cache\ItemInterface;
 
 #[CoversMethod(FlatMapper::class, 'createMapping')]
-#[CoversMethod(FlatMapper::class, 'createMappingRecursive')]
-#[CoversMethod(FlatMapper::class, 'getAttributeArgument')]
 #[CoversMethod(FlatMapper::class, 'setCacheService')]
+#[CoversMethod(FlatMapper::class, 'setValidateMapping')]
 #[CoversMethod(FlatMapper::class, 'setYamlMappings')]
+#[CoversMethod(MappingResolver::class, 'resolve')]
 #[CoversClass(FlatMapper::class)]
+#[CoversClass(MappingResolver::class)]
 #[CoversClass(MappingCreationException::class)]
 class FlatMapperCreateMappingTest extends TestCase
 {
@@ -52,14 +54,15 @@ class FlatMapperCreateMappingTest extends TestCase
     public function testCreateMappingWithCacheServiceDoesNotAssert(): void
     {
         $flatMapper = new FlatMapper();
+        $mappingResolver = new MappingResolver();
 
         // The intention is not to test the createMappingRecursive private method
         // but to dynamically give the CacheInterface mock a proper return value.
-        $reflectionMethod = (new \ReflectionClass(FlatMapper::class))->getMethod('createMappingRecursive');
+        $reflectionMethod = (new \ReflectionClass(MappingResolver::class))->getMethod('createMappingRecursive');
         $reflectionMethod->setAccessible(true);
         $cacheInterface = $this->createMock(CacheInterface::class);
         $cacheInterface->expects($this->once())->method('get')->willReturn(
-            $reflectionMethod->invoke($flatMapper, AuthorDTO::class)
+            $reflectionMethod->invoke($mappingResolver, AuthorDTO::class)
         );
 
         $flatMapper->setCacheService($cacheInterface);
@@ -158,6 +161,13 @@ class FlatMapperCreateMappingTest extends TestCase
         (new FlatMapper())->createMapping('ThisIsNotAValidClassString');
     }
 
+    public function testResolveWrongClassNameAsserts(): void
+    {
+        $this->expectException(MappingCreationException::class);
+        $this->expectExceptionMessageMatches("/An error occurred during mapping creation: ThisIsNotAValidClassString is not a valid class name/");
+        (new MappingResolver())->resolve('ThisIsNotAValidClassString');
+    }
+
     public function testCreateMappingWithSeveralIdenticalIdentifiersAsserts(): void
     {
         $this->expectException(MappingCreationException::class);
@@ -187,6 +197,15 @@ class FlatMapperCreateMappingTest extends TestCase
         $this->expectException(MappingCreationException::class);
         $this->expectExceptionMessageMatches("/cannot be readonly as it is non-scalar and/");
         (new FlatMapper())->createMapping(RootDTOWithReadonlyClassModifier::class);
+    }
+
+    public function testCreateMappingWithValidationDisabledDoesNotAssert(): void
+    {
+        $flatMapper = new FlatMapper();
+        $flatMapper->setValidateMapping(false);
+
+        $this->expectNotToPerformAssertions();
+        $flatMapper->createMapping(RootDTOWithNoIdentifier::class);
     }
 
     public function testCreateMappingWithReadonlyModifierOnScalarDtoSucceeds(): void
@@ -534,12 +553,12 @@ class FlatMapperCreateMappingTest extends TestCase
 
     public function testNormalizeYamlAttributeMapWithNullReturnsEmptyArray(): void
     {
-        $flatMapper = new FlatMapper();
-        $reflectionMethod = (new \ReflectionClass(FlatMapper::class))->getMethod('normalizeYamlAttributeMap');
+        $mappingResolver = new MappingResolver();
+        $reflectionMethod = (new \ReflectionClass(MappingResolver::class))->getMethod('normalizeYamlAttributeMap');
         $reflectionMethod->setAccessible(true);
 
         /** @var array<class-string, array<int|string, mixed>> $result */
-        $result = $reflectionMethod->invoke($flatMapper, null, 'class "Foo\\Bar\\Baz"');
+        $result = $reflectionMethod->invoke($mappingResolver, null, 'class "Foo\\Bar\\Baz"');
         $this->assertSame([], $result);
     }
 }
